+++ src/Reed/Classes/App/AppConstants.swift
@@ -49,7 +49,6 @@ enum AppConstants {
         static let windowState = "windowState"
         static let containerExpandedWindowState = "containerExpandedWindowState"
         static let readFeedsFilterState = "readFeedsFilterState"
-        static let readArticlesFilterState = "readArticlesFilterState"
         static let sidebarItemID = "sidebarItemID"
         static let feedIdentifier = "feedIdentifier" // legacy
     }